refactor: audit, cleanup, fixes, and readability

+++ contracts/src/interfaces/IAxelarGatewayMultisig.sol
@@ -5,15 +5,13 @@ pragma solidity >=0.8.0 <0.9.0;
 import { IAxelarGateway } from './IAxelarGateway.sol';
 
 interface IAxelarGatewayMultisig is IAxelarGateway {
-    event OwnershipTransferred(address[] preOwners, uint8 prevThreshold, address[] newOwners, uint8 newThreshold);
-    event OperatorshipTransferred(
-        address[] preOperators,
-        uint8 prevThreshold,
-        address[] newOperators,
-        uint8 newThreshold
-    );
+
+    event OwnershipTransferred(address[] preOwners, uint256 prevThreshold, address[] newOwners, uint256 newThreshold);
+
+    event OperatorshipTransferred(address[] preOperators, uint256 prevThreshold, address[] newOperators, uint256 newThreshold);
 
     function owners() external view returns (address[] memory);
 
     function operators() external view returns (address[] memory);
+
 }
